+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/extras/temperature.py
@@ -6,10 +6,16 @@ import sys
 import time
 import tempfile
 
-# so far this only supports the DHT22 sensor on either a
-# Raspberry Pi or an Orange Pi Zero 3
+# this module combines a number of different features to
+# have a consistent way of getting the temperature:
 #
-# for Raspberry Pi the pin number passed in is the BCM number,
+# - DHT22 sensor on either a Raspberry Pi or an Orange Pi Zero 3
+# - DHT11 sensor on (most?) Raspberry Pis
+# - BME280 sensor on (most?) Raspberry Pis
+# - TEMPer style USB sensors on any platform
+
+# for the gpio based sensors (DHT11 and DHT22)
+# on Raspberry Pi the pin number passed in is the BCM number,
 # so for GPIO4 (aka pin 7 on the 26 or 40 pin header) pass in 4
 # -- this means the DHT22 sensor is connected to
 #
@@ -85,6 +91,44 @@ class BME280_i2c:
         return data.temperature
 
 
+class USB_temper:
+    def __init__(self):
+        self.success = True
+        try:
+            import serial
+        except ImportError:
+            self.success, _ = run_subprocess("apt install -y python3-serial", timeout=600)
+            if not self.success:
+                logger.info("Failed to install python3-serial")
+                return
+        import temper
+
+        self.temper = temper.Temper()
+        known_devices = self.temper.list(use_json=True)
+        if known_devices is None:
+            self.success = False
+            logger.info("Did not find supported TEMPer USB devices")
+
+    def get_temperature(self):
+        if not self.success:
+            return None
+        info = self.temper.read()
+        if len(info) == 0:
+            logger.info("Failed to read temperature from USB sensor")
+            return None
+        for inf in info:
+            if "error" in inf:
+                logger.info(f"Failed to read temperature from USB sensor: {inf['error']}")
+            else:
+                # if there is an external probe - report that one as likely the
+                # user connected that in order to get a more accurate reading
+                if "external temperature" in inf:
+                    return inf["external temperature"]
+                if "internal temperature" in inf:
+                    return inf["internal temperature"]
+        return None
+
+
 class RPInative:
     def __init__(self, pin):
         self.pin = pin
@@ -144,6 +188,12 @@ class OPI:
         return None
 
 
+def usage(error_text=None):
+    if error_text:
+        print(f"ERROR: {error_text}")
+    print("Usage: temperature.py [pin|bme280|usb-temper]")
+
+
 if __name__ == "__main__":
     logger = logging.getLogger("temperature-service")
     logging.basicConfig(
@@ -155,16 +205,29 @@ if __name__ == "__main__":
     # command line arg - call with PIN number (BCM for RPi, PCxx for OPi Zero 3)
     pin = None
     use_bme280 = False
+    use_usb_temper = False
     # loop over the command line arguments
-    for arg in sys.argv:
+    for arg in sys.argv[1:]:
         if arg == "bme280":
+            if use_usb_temper:
+                usage("don't specify both bme280 and usb-temper")
+                sys.exit(1)
             use_bme280 = True
+        elif arg == "usb-temper":
+            if use_bme280:
+                usage("don't specify both bme280 and usb-temper")
+                sys.exit(1)
+            use_usb_temper = True
         else:
+            if use_bme280 or use_usb_temper:
+                usage("don't add a pin number with bme280 or usb-temper")
+                sys.exit(1)
             try:
                 pin = int(arg)
                 break
             except:
-                pass
+                usage("can't parse pin number")
+                sys.exit(1)
 
     if use_bme280:
         logger.info("Setting up i2c communication with BME280")
@@ -173,6 +236,14 @@ if __name__ == "__main__":
             logger.info("Initialized i2c communication with BME280")
         else:
             logger.error("Failed to initialize i2c communication with BME280")
+            sys.exit(1)
+    elif use_usb_temper:
+        logger.info("Attempting to use USB temperature sensor")
+        sensor = USB_temper()
+        if sensor.success:
+            logger.info("Initialized USB temperature sensor")
+        else:
+            logger.error("Failed to initialize USB temperature sensor")
             sys.exit(1)
     else:
         logger.info("Setting up communication with DHT11 or DHT22")